+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+## 2024.11.15
+
+### Added
+
+- Designed kernel implementation for DiscGolfScorecard component
+
+### Updated
+
+- Changed design to have the holes in a DiscGolfScorecard stored in a record instead of a static nested class
+
 ## 2024.10.31
 
 ### Added